Moving block templates into own folder
